app/requirements.txt
Flask
Flask-SQLAlchemy
pymysql
flask-marshmallow
flask-jwt-extended